+++ out/test/testFolder/gulpfile.js
@@ -1,0 +1,11 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const gulp_1 = require("gulp");
+function test() {
+    return;
+}
+exports.test = test;
+gulp_1.default.task("babel", () => {
+    return test();
+});
+//# sourceMappingURL=gulpfile.js.map