+++ test/testFolder/gulpfile.ts
@@ -1,0 +1,11 @@
+"use strict";
+
+import gulp from "gulp";
+
+export function test() {
+    return;
+}
+
+gulp.task("babel", () => {
+    return test();
+});