+++ test/testFolder/sub/gulpfile.js
@@ -1,0 +1,5 @@
+let gulp = require("gulp");
+
+gulp.task("sub", function() {
+    return;
+});